refactor description as its own react component
